refactor: move helper code to new dir
